Mise du code et des tests dans les bons fichiers
